+++ components/form/social-feed/standard-post-form.tsx
@@ -1,15 +1,20 @@
-"use client";
-
-import { Dispatch, SetStateAction, useEffect, useRef, useState } from "react";
+import React, {
+  Dispatch,
+  SetStateAction,
+  useEffect,
+  useRef,
+  useState,
+} from "react";
 import { useAuth } from "@clerk/nextjs";
 import { useMediaQuery } from "react-responsive";
-import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
-import { pollFormSchema, PollFormSchemaType } from "../types";
+import { zodResolver } from "@hookform/resolvers/zod";
 import { FeedInputButtons } from "./feed-input-buttons";
 import { useToast } from "@/app/hooks/use-toast";
-import { PollFields } from "@/components/form/social-feed/poll-fields";
-import { Textarea } from "@/components/shadcn/textarea";
+import {
+  standardPostFormSchema,
+  StandardPostFormSchemaType,
+} from "@/components/form/types";
 import {
   Form,
   FormControl,
@@ -17,10 +22,11 @@ import {
   FormItem,
   FormMessage,
 } from "@/components/shadcn/form";
+import { Textarea } from "@/components/shadcn/textarea";
 
 export type FeedInputMode = "POLL" | "STANDARD_POST";
 
-export function FeedInputPoll({
+export function StandardPostForm({
   feedInputMode,
   setFeedInputMode,
 }: {
@@ -34,35 +40,33 @@ export function FeedInputPoll({
   // TODO: Disable stuff if isLoading
   const [_isLoading, setIsLoading] = useState(false);
 
-  const pollForm = useForm<PollFormSchemaType>({
-    resolver: zodResolver(pollFormSchema),
+  const standardPostForm = useForm<StandardPostFormSchemaType>({
+    resolver: zodResolver(standardPostFormSchema),
     defaultValues: {
-      question: "",
-      options: [{ text: "" }, { text: "" }],
-      allowMultipleOptions: false,
+      content: "",
     },
   });
-  const question = pollForm.watch("question");
-  const textareaMaxLength = pollFormSchema.shape.question._def.checks.find(
-    (check) => check.kind === "max",
-  )?.value;
+  const content = standardPostForm.watch("content");
+  const textareaMaxLength =
+    standardPostFormSchema.shape.content._def.checks.find(
+      (check) => check.kind === "max",
+    )?.value;
 
   // Auto shrink and grow textarea
   const textareaRef = useRef<HTMLTextAreaElement>(null);
   const textareaMaxHeight = 300;
   const textareaMinHeight = 48;
   const placeholder = isSmallScreen
-    ? "Ask something"
-    : "What do you wanna ask to the community?";
+    ? "Say something"
+    : "Don't be shy, say something!";
   useEffect(() => {
     const textarea = textareaRef.current;
     if (!textarea) return;
     textarea.style.height = `${textareaMinHeight}px`;
     textarea.style.height = `${textarea.scrollHeight}px`;
-  }, [question]);
+  }, [content]);
 
-  // Functions
-  const onSubmitPoll = async (values: PollFormSchemaType) => {
+  const onSubmitStandardPost = async (values: StandardPostFormSchemaType) => {
     try {
       setIsLoading(true);
       const token = await getToken();
@@ -73,17 +77,17 @@ export function FeedInputPoll({
       //TODO: Plug endpoint here
       console.log(values);
 
-      pollForm.reset();
+      standardPostForm.reset();
       toast({
         // title: t("toast-creation-success"),
-        title: "TODO: trad (Poll creation success)",
+        title: "TODO: trad (Post creation success)",
       });
       // router.push(`/polls`);
     } catch (err) {
       toast({
         variant: "destructive",
         // title: t("toast-creation-error"),
-        title: "TODO: trad (Poll creation error)",
+        title: "TODO: trad (Post creation error)",
       });
       console.error("error", err);
     }
@@ -91,26 +95,23 @@ export function FeedInputPoll({
   };
 
   return (
-    <Form {...pollForm}>
+    <Form {...standardPostForm}>
       <form
-        onSubmit={pollForm.handleSubmit(onSubmitPoll)}
+        onSubmit={standardPostForm.handleSubmit(onSubmitStandardPost)}
         className="flex flex-col gap-4"
       >
         <div className="flex flex-row gap-4">
           <FormField
             rules={{ required: true }}
-            control={pollForm.control}
-            name="question"
+            control={standardPostForm.control}
+            name="content"
             render={({ field }) => (
               <FormItem className="relative w-full">
                 <FormControl>
                   <Textarea
                     ref={textareaRef}
                     onChange={(e) => field.onChange(e.target.value)}
-                    style={{
-                      minHeight: textareaMinHeight,
-                      maxHeight: textareaMaxHeight,
-                    }}
+                    className={`!min-h-[${textareaMinHeight}px] !max-h-[${textareaMaxHeight}px]`}
                     placeholder={placeholder}
                     maxLength={textareaMaxLength}
                   />
@@ -125,8 +126,6 @@ export function FeedInputPoll({
             setFeedInputMode={setFeedInputMode}
           />
         </div>
-
-        <PollFields form={pollForm} />
       </form>
     </Form>
   );
